1. Added BottomUpMergeSort 2. Cleanup

+++ src/main/java/algorithms/inversions/counter/Counter.java
@@ -2,6 +2,6 @@ package algorithms.inversions.counter;
 
 public interface Counter<T extends Comparable<T>> {
 
-	int count(T[] a);
+	long count(T[] a);
 
 }
